requirements.txt
datasets
sentence_transformers
torch==2.3.1
scipy
pandas
matplotlib
umap-learn
